requirements.txt
flask==2.3.3
line-bot-sdk==3.12.2
requests==2.31.0
gunicorn==21.2.0
